CNDB-6523 (CASSANDRA-18311) fix BufferPool.memoryInUse counter (#602)
The counter was incorrectly decremented by the size of
the unused portion of the provided buffer.
It is now decremented by the number of bytes actually
returned to the pool (that may be different than "size").
The number is calculated as a difference between original
and resulting buffer capacity.

(cherry picked from commit cbea20addbfa41d3ccb448271ccfa9ac48aab4c7)